+++ src/avt/Expressions/Math/avtExpExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                               avtExpExpression.C                              //
+//                               avtExpExpression.C                          //
 // ************************************************************************* //
 
 #include <avtExpExpression.h>
@@ -110,7 +110,7 @@ avtExpExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, exp(val));
         }
     }
