+++ src/avt/Expressions/Math/avtRadianToDegreeExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                          avtRadianToDegreeExpression.C                        //
+//                          avtRadianToDegreeExpression.C                    //
 // ************************************************************************* //
 
 #include <avtRadianToDegreeExpression.h>
@@ -115,7 +115,7 @@ avtRadianToDegreeExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, val*(360./(2.*M_PI)));
         }
     }
