+++ src/avt/Expressions/Math/avtRoundExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                               avtRoundExpression.C                            //
+//                               avtRoundExpression.C                        //
 // ************************************************************************* //
 
 #include <avtRoundExpression.h>
@@ -110,7 +110,7 @@ avtRoundExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, (val < 0.) ?
                               -floor(fabs(val)+0.5) : floor(val+0.5));
         }
