+++ src/avt/Expressions/Math/avtTanhExpression.C
@@ -110,7 +110,7 @@ avtTanhExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, tanh(val));
         }
     }
